+++ ft_strlen.c
@@ -6,7 +6,7 @@
 /*   By: zalouach <zalouach@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/11/01 21:01:54 by zalouach          #+#    #+#             */
-/*   Updated: 2023/12/01 19:45:20 by zalouach         ###   ########.fr       */
+/*   Updated: 2023/12/12 17:48:29 by zalouach         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,7 +15,7 @@
 size_t	ft_strlen(const char *s)
 
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while (s[i])
